后端的第二次尝试

+++ wxprojBack/src/main/java/com/tickoff/domain/User.java
@@ -5,13 +5,12 @@ import lombok.ToString;
 
 @Data
 @ToString
-
 public class User {
-    private int id;
+    private Long id;
     private String username;
     private String password;
     /* 省略get、set和带参构造方法 */
-    public User(int id,String username,String password){
+    public User(Long id,String username,String password){
         this.id=id;
         this.username=username;
         this.password=password;
